+++ src/gen/templates/worker_test.t
@@ -14,14 +14,13 @@ use loco_rs::testing;
 
 use blo::workers::{{module_name}}::{{struct_name}}Worker;
 use blo::workers::{{module_name}}::{{struct_name}}WorkerArgs;
-use migration::Migrator;
 use serial_test::serial;
 
 
 #[tokio::test]
 #[serial]
 async fn test_run_{{module_name}}_worker() {
-    let boot = testing::boot_test::<App, Migrator>().await.unwrap();
+    let boot = testing::boot_test::<App>().await.unwrap();
 
     // Execute the worker ensuring that it operates in 'ForegroundBlocking' mode, which prevents the addition of your worker to the background
     assert!(